refactored modals to no have duplicate code
