+++ components/ui/CreateResultModal.tsx
